+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Change Log
 - [FIXED] a problem with create_nxgraph
 - [ADDED] parameter slack_as_gen for `from_ucte()` converter (changed the default behavior)
 - [FIXED] cim2pp: the `ignore_errors` parameter was not respected in the `CimParser`
+- [FIXED] cim2pp: keep the busbar names and IDs
 
 [3.0.0] - 2025-03-06
 -------------------------------